Moved activities to an activites folder
