+++ TaskManagerAndroid/TaskManagerAndroid.csproj
@@ -60,12 +60,12 @@
     <Reference Include="Mono.Android" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="LoginActvitiy.cs" />
-    <Compile Include="MainActivity.cs" />
-    <Compile Include="ProjectCreateActivity.cs" />
-    <Compile Include="ProjectListActivity.cs" />
-    <Compile Include="ProjectOptionsActivity.cs" />
-    <Compile Include="RegisterActivity.cs" />
+    <Compile Include="Activities\LoginActvitiy.cs" />
+    <Compile Include="Activities\MainActivity.cs" />
+    <Compile Include="Activities\ProjectCreateActivity.cs" />
+    <Compile Include="Activities\ProjectListActivity.cs" />
+    <Compile Include="Activities\ProjectOptionsActivity.cs" />
+    <Compile Include="Activities\RegisterActivity.cs" />
     <Compile Include="Resources\Resource.Designer.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Token.cs" />
